Fixed two warnings

+++ parser.h
@@ -99,7 +99,7 @@ public:
 						
 					// Add the rule in reverse
 					vector<symbol*> rhsRev = rule.rhsReversed();
-					for (int i = 0; i < rhsRev.size(); ++i) {
+					for (size_t i = 0; i < rhsRev.size(); ++i) {
 						workingStack.push(rhsRev[i]);
 					}
 				} 
@@ -135,7 +135,7 @@ public:
 			cout << current->getSymbol()->getName() + " >>> ";
 
 		}
-		for (int i = 0; i < current->getChildren()->size(); ++i)//--------------------------for(auto it = current->getChildren().begin(); it != current->getChildren().begin(); ++it)
+		for (size_t i = 0; i < current->getChildren()->size(); ++i)//--------------------------for(auto it = current->getChildren().begin(); it != current->getChildren().begin(); ++it)
 		{
 			cout << current->getSymbol()->getName() << " >>> ";
 			printPST(current->getChildren()->at(i),root);
